M5L1: fix some MR comments.
