+++ ok-marketplace-biz/src/commonTest/kotlin/repo/BizRepoReadTest.kt
@@ -1,4 +1,4 @@
-package ru.otus.otuskotlin.marketplace.biz.repo
+package repo
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
