+++ ok-marketplace-biz/src/commonTest/kotlin/validation/ValidationBadTitle.kt
@@ -1,4 +1,4 @@
-package ru.otus.otuskotlin.marketplace.biz.validation
+package validation
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
